Fix -Wshorten-64-to-32 for musl
On musl uint_fast32_t is uint32_t, while for glibc it is unsigned long
long.

+++ src/Common/PoolWithFailoverBase.h
@@ -353,7 +353,7 @@ struct PoolWithFailoverBase<TNestedPool>::PoolState
     }
 
 private:
-    std::minstd_rand rng = std::minstd_rand(randomSeed());
+    std::minstd_rand rng = std::minstd_rand(static_cast<uint_fast32_t>(randomSeed()));
 };
 
 template <typename TNestedPool>
